Sliders and their connection with other classes have been improved
Slidermediator was converted, now it is not abstract. Created mediators: MusicSliderAudioMixerTunerMediator, SoundSliderAudioMixerTunerMediator, SensitivitySliderInputAxisControllerMediator and GamePauserAudioMixerTunerMediator, For the connection of values ​​in the slider models with other systems. Simplified by AudioMixerTuner. The registration of these mediators in LifetimeScopes changed. Game Settings assembly definition has been created, all settings are moved into it.